+++ src/default/default_amber.js
@@ -4,15 +4,15 @@
 
     thisModule.config(config);
 
-    function config($mdThemingProvider, pipTranslateProvider) {
-        pipTranslateProvider.translations('en', {
-            THEME: 'Theme',
-            amber: 'Amber'
-        });
-        pipTranslateProvider.translations('ru', {
-            THEME: 'Тема',
-            amber: 'Янтарная'
-        });
+    function config($mdThemingProvider) {
+        // pipTranslateProvider.translations('en', {
+        //     THEME: 'Theme',
+        //     amber: 'Amber'
+        // });
+        // pipTranslateProvider.translations('ru', {
+        //     THEME: 'Тема',
+        //     amber: 'Янтарная'
+        // });
 
         var orangeBackgroundPalette = $mdThemingProvider.extendPalette('grey', {
             'A100': 'rgba(250, 250, 250, 1)',