+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Pip.WebUI.Themes  Changelog
 
-## <a name="1.0.0"></a> 1.0.0 (2016-07-30)
+## <a name="1.5.0"></a> 1.5.0 (2016-10-15)
 
 Initial release with core services, behaviors and switching color themes
 
 #### Features
 * **themes** Themes switching services
+* **default** Default themes
+* **bootbarn** BootBarn themes
 
 #### Breaking Changes
 No breaking changes in this version